+++ docs/changes.md
@@ -1,7 +1,8 @@
 ## Main changes
 
 - added gpu acceleration for CLI volumetric visualization in #832 by @zhang-qiang-github
-
+- fixes for `k3d`
+- added `plotter.fov(value)` (field of view)
 
 
 -------------------------